requirements.txt
tk		8.6.12
pandas	1.4.4	
numpy		1.21.5
json5		0.9.6
sqlalchemy	1.4.39
PyMySQL	1.0.2
requests	2.28.2
python-dateutil	2.8.2
pillow	9.2.0	